requirements.txt
pygame
pyopengl
pillow